docs: convert remaining examples
