+++ site/examples/plaintext.tsx
@@ -1,13 +1,14 @@
 import React, { useState, useMemo } from 'react'
-import { createEditor } from 'slate'
+import { Node, createEditor } from 'slate'
 import { Slate, Editable, withReact } from 'slate-react'
+import { withHistory } from 'slate-history'
 
-const ReadOnlyExample = () => {
-  const [value, setValue] = useState(initialValue)
-  const editor = useMemo(() => withReact(createEditor()), [])
+const PlainTextExample = () => {
+  const [value, setValue] = useState<Node[]>(initialValue)
+  const editor = useMemo(() => withHistory(withReact(createEditor())), [])
   return (
     <Slate editor={editor} value={value} onChange={value => setValue(value)}>
-      <Editable readOnly placeholder="Enter some plain text..." />
+      <Editable placeholder="Enter some plain text..." />
     </Slate>
   )
 }
@@ -20,4 +21,4 @@ const initialValue = [
   },
 ]
 
-export default ReadOnlyExample
+export default PlainTextExample
