+++ site/examples/read-only.tsx
@@ -1,14 +1,13 @@
 import React, { useState, useMemo } from 'react'
-import { createEditor } from 'slate'
+import { createEditor, Node } from 'slate'
 import { Slate, Editable, withReact } from 'slate-react'
-import { withHistory } from 'slate-history'
 
-const PlainTextExample = () => {
-  const [value, setValue] = useState(initialValue)
-  const editor = useMemo(() => withHistory(withReact(createEditor())), [])
+const ReadOnlyExample = () => {
+  const [value, setValue] = useState<Node[]>(initialValue)
+  const editor = useMemo(() => withReact(createEditor()), [])
   return (
     <Slate editor={editor} value={value} onChange={value => setValue(value)}>
-      <Editable placeholder="Enter some plain text..." />
+      <Editable readOnly placeholder="Enter some plain text..." />
     </Slate>
   )
 }
@@ -21,4 +20,4 @@ const initialValue = [
   },
 ]
 
-export default PlainTextExample
+export default ReadOnlyExample
